moved relevant states into Goals component
